// src/app/api/model/payment.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Invoice } from './invoice';


export interface Payment { 
    id?: string;
    invoiceId?: string | null;
    amount?: number;
    paymentDate?: string;
    method?: string | null;
    reference?: string | null;
    createdAt?: string | null;
    invoice?: Invoice;
}