// src/app/api/model/paymentRequest.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentRequest { 
    invoiceId?: string;
    amount?: number;
    paymentDate?: string;
    method?: string | null;
    reference?: string | null;
}